.github/requirements.txt
zhipuai>=2.1.0
PyGithub>=2.3.0
